Expanded world map mk. 1
The first version of the bigger world map (titled “Seagull_Island_v2”)
has been created and exported to an .fbx in the working files folder.
This version currently consists of only the main components, and also
has an issue with some non-manifold geometry. It will be reworked
slightly, and built upon.