Item Spawns
This looks like a big update, but only because I had to put the prefabs
into the "Resources" folder in order to iteratively refer to each prefab
in code.

We can set spawn times and locations for different prefabs and they will
appear at that time in those places. This could potentially be
randomized in both time, location, and quantity of items.

Code will not spawn items currently, though because we should put things
into the world in order to see how it will all look before we jump to
actually spawning. (If you wantto check out with spawning, just remove
the comment slashes around "SpawnPickups" in update, and edit the prefab
you want to have a spawn time and spawn position.