FIXED THE DAMN BUG
ok So I FINALLY found out where the bug was coming from that stops the
items. The player prefags were all untagged objects, so sometimes as you
collided with the beam, you also collide with the player, or it's
identifier. The object drops because it hits a nest, then immediantly
stops in the air because it hit an untagged thing. Players are tagged
player now